fix(providers): align embedding usage metadata
  - parse Bedrock Cohere usage from input-token-count headers
  - preserve NaN embedding tokens as JSON null when required
  - detect Cohere cross-region embedding profile IDs
  - map OpenAI-compatible embedding usage from prompt tokens
  - normalize embedding response metadata across providers

+++ pkg/providers/alibaba/embedding_model.go
@@ -137,7 +137,7 @@ func (m *EmbeddingModel) DoEmbedMany(ctx context.Context, inputs []string, opts 
 	return &types.EmbeddingsResult{
 		Embeddings: embeddings,
 		Usage: types.EmbeddingUsage{
-			Tokens:      inputTokens,
+			Tokens:      float64(inputTokens),
 			InputTokens: inputTokens,
 			TotalTokens: inputTokens,
 		},
